requirements.txt
# 项目依赖
httpx>=0.27.0
python-dotenv>=1.0.0